+++ Source/Atum/Private/Layers/Activation/AtumLayerTanh.cpp
@@ -19,7 +19,7 @@ bool UAtumLayerTanh::OnForward_Implementation(
 )
 {
 	Output = DuplicateObject(Input.GetObject(), nullptr);
-	Output->SetData((*Module)(Input->GetDataChecked().to(c10::kDouble)));
+	Output->SetData((*Module)(Input->GetDataChecked()));
 	return true;
 }
